app/requirements.txt
streamlit
mysql-connector-python
faker
python-dotenv
pandas
plotly
